get rid of redirect manager
