+++ react/index.js
@@ -1,15 +1,39 @@
-import React from 'react'
+import React, { useEffect } from 'react'
+import queryString from 'query-string'
+import { head, path } from 'ramda'
 
-import AddressOrderFormProvider from './components/AddressOrderFormProvider'
-import RedirectManager from './components/RedirectManager'
+import { useAddress, withAddressProvider } from './components/AddressContext'
+import AddressPage from './components/AddressPage'
 
 import './global.css'
 
-const AddressManager = props => (
-  <AddressOrderFormProvider>
-    <RedirectManager { ...props } />
-  </AddressOrderFormProvider>
-)
+const redirectToReturnURL = () => {
+  try {
+    const parsedQueryString = queryString.parse(window.location.search)
+    const returnURL = parsedQueryString && parsedQueryString.returnUrl  || ''
+    const cleanUrl = head(returnURL) === '/' ? returnURL : `/${returnURL}`
+    window.location.href = cleanUrl
+  } catch (e) {
+    // Unable to redirect
+  }
+}
+
+const AddressManager = props => {
+  const { address } = useAddress()
+
+  const checkIfAddressIsSet = () => {
+    if (!!path(['orderForm', 'shippingData', 'address'], address)) {
+      redirectToReturnURL()
+    }
+  }
+
+  useEffect(() => {
+    checkIfAddressIsSet()
+  })
+  return (
+    <AddressPage {...props} onSelectAddress={redirectToReturnURL} />
+  )
+}
 
 AddressManager.schema = {
   title: 'address-locator.address-manager-title',
@@ -26,4 +50,4 @@ AddressManager.schema = {
   }
 }
 
-export default AddressManager
+export default withAddressProvider(AddressManager)
